[ing]채팅방

+++ src/main/java/com/realEstate/realEstate/model/dto/chatting/AggregationDto.java
@@ -8,7 +8,7 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class AggregationDto implements Serializable {
 
-    private Integer saleNo;
+    private Long saleNo;
     private String isIncrease;
     private AggregationTarget target;
 }
